--- code-challenges/301/arrayReverse.1/ReadMe.md
@@ -1,13 +1,0 @@
-# Reverse an Array
-This JavaScript file includes a method that takes in an array and returns a reversed array
-
-## Challenge
-Take an array and reverse it. For instance an array of [1,2,3,4] would then equal [4,3,2,1]
-
-## Approach & Efficiency
-The approach was to use a for loop and unshift each element to a new array which would reverse the array.
-This Big O space would be O(n) since the the amount of space is in relation to the size of the array.
-the Big O time would be O(n) since there is one operation on each index so the bigger the array, the bigger n.
-
-## Solution
-![(../assets/reverseArray.jpg)]
