--- code-challenges/301/arrayReverse/array-reverse.js
@@ -1,14 +1,0 @@
-'use strict'
-
-let arr =[1,2,3,4,5]
-
-
-function reverseArray(arr){
-
-    let result = [];
-    for(let i =0; i<arr.length; i++){
-
-        result.unshift(arr[i]);
-    }
-    return result;
-}
