--- code-challenges/301/array_shift/README.md
@@ -1,13 +1,0 @@
-# Insert and shift middle index of array
-This is a function that takes in an arrray and finds the middle, it then inserts the passed value into where the middle index is and shifts all indexes.
-
-## Challenge
-The challenge was to find a way to insert a new item in the middle of the array while keeping the same order.
-
-## Approach & Efficiency
-We decided to find the middle of the array and loop from mid to end and shift all of those elements.
-space complexity is O(1)
-time complexity is O(n) n being the size of the array.
-
-## Solution
-![(../assets/array_shift.jpg)]
