--- code-challenges/301/jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  verbose: true,
-};
