--- code-challenges/301/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "401-js-proof-of-life",
-  "version": "1.0.0",
-  "description": "Proof of Life",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "lint": "eslint **/*.js",
-    "test": "jest --verbose --coverage",
-    "test-watch": "jest --watchAll --verbose --coverage"
-  },
-  "author": "John Cokos",
-  "license": "MIT",
-  "dependencies": {
-    "eslint": "^5.6.1",
-    "faker": "^4.1.0",
-    "jest": "^23.6.0",
-    "lint": "^1.1.2",
-    "travis": "^0.1.1"
-  }
-}
