+++ code-challenges/Array/array_shift/README.md
@@ -10,4 +10,4 @@ space complexity is O(1)
 time complexity is O(n) n being the size of the array.
 
 ## Solution
-![(../assets/array_shift.jpg)]
+![(../assets/array_binary_search)]
